--- Dev/raw.dat
@@ -1,3 +1,0 @@
-2             1             
-4             2             
-6             3
