--- Dev/structured.dat
@@ -1,10 +1,0 @@
-2             1             
-2.44444       1.22222       
-2.88889       1.44444       
-3.33333       1.66667       
-3.77778       1.88889       
-4.22222       2.11111       
-4.66667       2.33333       
-5.11111       2.55556       
-5.55556       2.77778       
-6             3
