+++ Source/Utilities/CD_LookupTable1D.H
@@ -5,7 +5,7 @@
 
 /*!
   @file   CD_LookupTable1D.H
-  @brief  Declaration of a table for looking up coefficients etc. 
+  @brief  Declaration of a lookup table in one independent variable.
   @author Robert Marskar
 */
 
@@ -16,220 +16,220 @@
 #include <iostream>
 #include <vector>
 #include <array>
-#include <tuple>
-
-// Chombo includes
-#include <REAL.H>
+#include <type_traits>
 
 // Our includes
-#include <CD_NamespaceHeader.H>
 #include <CD_LookupTable.H>
 
 /*!
-  @brief Class for interpolation of f = f(x) data. 
+  @brief Class for interpolation of f = f(x) data in one independent variable x.
+  @details N is the number of additional dependent variables. T must be a floating point type (e.g., float, double)
 */
-template <int N>
+template <typename T, size_t N = 1, typename I = std::enable_if_t<std::is_floating_point<T>::value>>
 class LookupTable1D
 {
 public:
-  static_assert(N > 1, "LookupTable1D<N> must have N > 1");
+  static_assert(N >= 1, "LookupTable1D<N> must have N >= 1");
+
   /*!
     @brief Default constructor. Creates a table without any entries. 
   */
-  LookupTable1D();
-
-  /*!
-    @brief Copy constructor. 
-  */
-  LookupTable1D(const LookupTable1D& a_table);
+  LookupTable1D() noexcept;
 
   /*!
     @brief Destructor (does nothing). 
   */
-  virtual ~LookupTable1D();
+  virtual ~LookupTable1D() noexcept = default;
 
   /*!
-    @brief Clear all data
+    @brief Reset everything.
   */
   inline void
-  clear();
-
-  /*!
-    @brief Make the table into a uniform table with the number of entries. 
-    @details This will make the data uniform with constant spacing along the specified column data. 
-    @note The spacing that is used is based on the smallest and largest value in the specified column. 
-  */
-  inline void
-  makeUniform(const int a_numRows);
-
-  /*!
-    @brief Scale data in column by a specified factor. 
-    @details The template parameter is the column to be scaled. 
-    @param[in] a_scale       Factor to scale by
-    @param[in] a_whichColumn Column to scale
-  */
-  template <int K>
-  inline void
-  scale(const Real a_scale);
-
-  /*!
-    @brief Swap two columns
-    @param[in] a_column1 First column
-    @param[in] a_column2 Second column
-  */
-  inline void
-  swap(const int a_column1, const int a_column2);
-
-  /*!
-    @brief Restrict table range to maximum and minimum. Data outside these ranges are discarded. 
-    @param[in] a_min         Minimum x-value in the table. 
-    @param[in] a_max         Maximum x-value in the table. 
-    @param[in] a_whichColumn Ranges based on this column. 
-  */
-  inline void
-  setRange(const Real a_min, const Real a_max, const int a_whichColumn = 0);
-
-  /*!
-    @brief Restrict table range to minimum value. Data outside these ranges are discarded. 
-    @param[in] a_min         Minimum x-value in the table. 
-    @param[in] a_whichColumn Ranges based on this column. 
-  */
-  inline void
-  setMinRange(const Real a_min, const int a_whichColumn = 0);
-
-  /*!
-    @brief Restrict table range to maximum value. Data outside this range are discarded. 
-    @param[in] a_max         Maximum x-value in the table. 
-    @param[in] a_whichColumn Ranges based on this column. 
-  */
-  inline void
-  setMaxRange(const Real a_max, const int a_whichColumn = 0);
-
-  /*!
-    @brief Set the strategy when retrieving data that is out of range on the low side
-    @details Set the input parameter a_strategy to Constant if you want to virtually pad the data with the last endpoint or Linear if you want linear interpolation.
-    @param[in] a_strategy Range strategy. 
-  */
-  inline void
-  setOutOfRangeStrategyLow(const OutOfRangeStrategy a_strategy);
-
-  /*!
-    @brief Set the strategy when retrieving data that is out of range on the high side
-    @details Set the input parameter a_strategy to Constant if you want to virtually pad the data with the last endpoint or Linear if you want linear interpolation.
-    @param[in] a_strategy Range strategy. 
-  */
-  inline void
-  setOutOfRangeStrategyHigh(const OutOfRangeStrategy a_strategy);
-
-  /*!
-    @brief Set the table spacing mode
-    @param[in] a_spacing Spacing mode.
-  */
-  inline void
-  setTableSpacing(const TableSpacing a_spacing);
-
-  /*!
-    @brief Sort the table. Useful if the input data was not already sorted. 
-    @param[in] a_whichColumn Sort along this column. 
-    @details On output, the data in the table is sorted -- smallest x goes first. 
-  */
-  inline void
-  sort(const int a_whichColumn = 0);
+  reset() noexcept;
 
   /*!
     @brief Add entry. 
-    @param[in] x Entry to add. You can add an entire array through the variadic template. 
+    @param[in] x Entry to add. For example addData(1,1,1,1). Number of elements in x must be N+1
   */
   template <typename... Ts>
   inline void
-  addEntry(const Ts&... x);
+  addData(const Ts&... x);
 
   /*!
-    @brief Dump table to pout()
+    @brief Utility function for swapping columns. 
+    @details This is done on the raw data -- if the user wants to swap columns in the resampled/structured data
+    then he needs to call this function first and then resample the table through prepareTable
+    @param[in] a_columnOne Column to swap.
+    @param[in] a_columnTwo Column to swap.
   */
   inline void
-  dumpTable(std::ostream& a_outputStream = std::cout) const;
+  swap(const size_t a_columnOne, const size_t a_columnTwo) noexcept;
 
   /*!
-    @brief Dump table to a file. 
-    @param[in] a_fileName File name
+    @brief Utility function which scales one of the columns (either dependent or independent variable)
+  */
+  template <size_t K>
+  inline void
+  scale(const T& a_scale) noexcept;
+
+  /*!
+    @brief Utility function for truncating raw data along one of the variables (either dependent or independent).
+    @details This will discard (from the raw data) all data that fall outside the input interval. This is done on the raw data -- the user will
+    need to call prepareTable if the result should propagate into the resampled/structured data.
+    @param[in] a_min Minimum value represented.
+    @param[in] a_max Maximum value represented.
   */
   inline void
-  dumpTable(const std::string a_fileName) const;
+  truncate(const T& a_min, const T& a_max, const size_t a_column) noexcept;
 
   /*!
-    @brief Get entry
-    @note When calling this, make sure the data is sorted AND uniform. For performance reasons this is not checked when calling this routine. 
-    @param[in] a_x Interpolation point (independent variable)
+    @brief Set the out-of-range strategy on the low end
+    @param[in] a_strategy Out-of-range strategy on low end
   */
-  template <int K>
-  inline Real
-  interpolate(const Real a_x) const;
+  inline void
+  setRangeStrategyLo(const OutOfRangeStrategy& a_strategy) noexcept;
 
   /*!
-    @brief Get data.
-    @note When calling this, make sure the data is sorted AND uniform. For performance reasons this is not checked when calling this routine. 
-    @param[in] a_x Interpolation point (independent variable)
+    @brief Set the out-of-range strategy on the high end
+    @param[in] a_strategy Out-of-range strategy on the high end
   */
-  inline std::array<Real, N>
-  interpolate(const Real a_x) const;
+  inline void
+  setRangeStrategyHi(const OutOfRangeStrategy& a_strategy) noexcept;
 
   /*!
-    @brief Get size of the table.
-    @returns m_data.size()
+    @brief Turn the raw data into uniform data for fast lookup.
+    @param[in] a_independentVariable The independent variable (i.e., column in the input data).
+    @param[in] a_numPoints Number of points in the input table. 
+    @param[in] a_coordSys Coordinate system spacing. 
   */
-  inline int
-  getNumEntries() const;
+  inline void
+  prepareTable(const size_t& a_independentVariable, const size_t& a_numPoints, const CoordinateSystem& a_coordSys);
+
+  /*!
+    @brief Interpolation function for specific dependent variable K
+    @param[in] a_x Independent variable x
+  */
+  template <size_t K>
+  inline T
+  interpolate(const T& a_x) const;
+
+  /*!
+    @brief Interpolate whole table.
+    @param[in] a_x Independent variable x
+  */
+  inline std::array<T, N + 1>
+  interpolate(const T& x) const;
+
+  /*!
+    @brief Access function for raw data. 
+    @return Returns m_rawData
+  */
+  inline std::vector<std::array<T, N + 1>>&
+  getRawData() noexcept;
+
+  /*!
+    @brief Access function for structured data. 
+    @return Returns m_structuredData
+  */
+  inline std::vector<std::array<T, N + 1>>&
+  getStructuredData() noexcept;
+
+  /*!
+    @brief Access function for raw data.
+    @return Returns m_rawData
+  */
+  inline const std::vector<std::array<T, N + 1>>&
+  getRawData() const noexcept;
+
+  /*!
+    @brief Access function for structured data.
+    @return Returns m_structuredData
+  */
+  inline const std::vector<std::array<T, N + 1>>&
+  getStructuredData() const noexcept;
+
+  /*!
+    @brief Dump raw table data to file
+    @param[in] a_file File name
+  */
+  inline void
+  writeRawData(const std::string& a_file) const noexcept;
+
+  /*!
+    @brief Dump structured table data to file
+    @param[in] a_file File name
+  */
+  inline void
+  writeStructuredData(const std::string& a_file) const noexcept;
+
+  /*!
+    @brief Dump raw table data to output stream.
+    @param[in] a_ostream Output stream
+  */
+  inline void
+  outputRawData(std::ostream& a_ostream = std::cout) const noexcept;
+
+  /*!
+    @brief Dump structured table data to file.
+    @param[in] a_ostream Output stream
+  */
+  inline void
+  outputStructuredData(std::ostream& a_ostream = std::cout) const noexcept;
 
 protected:
   /*!
-    @brief End point stategy on the low end of the data 
+    @brief Check if data can be interpolated
   */
-  OutOfRangeStrategy m_strategyLo;
+  bool m_isGood;
 
   /*!
-    @brief End point stategy on the high end of the data 
+    @brief Out-of-range strategy on low end
   */
-  OutOfRangeStrategy m_strategyHi;
+  OutOfRangeStrategy m_rangeStrategyLo;
 
   /*!
-    @brief For setting the table spacing
+    @brief Out-of-range strategy on high end
   */
-  TableSpacing m_spacing;
+  OutOfRangeStrategy m_rangeStrategyHi;
 
   /*!
-    @brief Table state. First entry is the column along which we sort and make the table uniform. The second entry is 
-    whether or not the table is actually sorted, and the third entry is whether or not the table is uniform along the sorted column.
+    @brief Underlying 1D grid. This is populated when calling prepareTable.
   */
-  std::tuple<int, bool, bool> m_state;
+  std::tuple<CoordinateSystem, size_t, T, T, T> m_grid;
 
   /*!
-    @brief Table resolution.
+    @brief Raw data
   */
-  Real m_delta;
+  std::vector<std::array<T, N + 1>> m_rawData;
 
   /*!
-    @brief Data
+    @brief Structured data. This is populated when calling prepareTable.
   */
-  std::vector<std::array<Real, N>> m_data;
+  std::vector<std::array<T, N + 1>> m_structuredData;
 
   /*!
-    @brief Direct lookup, assumes non-uniform ordering. 
-    @details This is very slow, because it will look through the entire table. 
-    @note The data MUST be sorted along the input column for this to work. 
+    @brief Utility function for outputting data.
+    @param[in] a_ostream Output stream
+    @param[in] a_data Data to be sent to output stream
   */
-  inline std::array<Real, N>
-  directLookup(const Real a_x) const;
+  inline void
+  outputData(std::ostream& a_ostream, const std::vector<std::array<T, N + 1>>& a_data) const noexcept;
 
   /*!
-    @brief Get the lower index during a lookup
-    @param[in] a_x Interpolation point. 
+    @brief Utility function for outputting data to a file
+    @param[in] a_file File name
+    @param[in] a_data Data to be sent to output stream
   */
-  inline int
-  getIndexLo(const Real a_x) const;
+  inline void
+  writeToFile(const std::string& a_file, const std::vector<std::array<T, N + 1>>& a_data) const noexcept;
+
+  /*!
+    @brief Get the lower index that brackets the input variable between two data points in the structured grid.
+    @param[in] a_x Independent variable.
+  */
+  inline size_t
+  getIndexLo(const T& a_x) const;
 };
-
-#include <CD_NamespaceFooter.H>
 
 #include <CD_LookupTable1DImplem.H>
